+++ src/index.ts
@@ -1,0 +1,3 @@
+import JSONCache from './lib/jsonCache';
+
+export default JSONCache;